task5-exceptions

+++ src/main/java/com/stackroute/service/MovieServiceImpl.java
@@ -1,9 +1,9 @@
 package com.stackroute.service;
 
 import com.stackroute.domain.Movie;
-/*import com.stackroute.exceptions.DatabaseTemporarilyUnavailableException;
+import com.stackroute.exceptions.DatabaseTemporarilyUnavailableException;
 import com.stackroute.exceptions.MovieAlreadyExistsException;
-import com.stackroute.exceptions.MovieDoesNotExistException;*/
+import com.stackroute.exceptions.MovieDoesNotExistException;
 import com.stackroute.repository.MovieRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,34 +20,34 @@ public class MovieServiceImpl implements MovieService {
     }
 
     @Override
-    public Movie saveMovie(Movie movie) /*throws MovieAlreadyExistsException */{
-      /*  if (movieRepository.existsById(movie.getId())) {
+    public Movie saveMovie(Movie movie) throws MovieAlreadyExistsException {
+       if (movieRepository.existsById(movie.getId())) {
             throw new MovieAlreadyExistsException("Movie Already Exists");
-        }*/
+        }
         Movie savedMovie = movieRepository.save(movie);
 
         return savedMovie;
     }
 
     @Override
-    public List<Movie> getAllMovies()/* throws DatabaseTemporarilyUnavailableException */{
-       /* if (movieRepository.findAll() == null) throw new DatabaseTemporarilyUnavailableException("Database Temporarily Unavailable");*/
+    public List<Movie> getAllMovies() throws DatabaseTemporarilyUnavailableException {
+        if (movieRepository.findAll() == null) throw new DatabaseTemporarilyUnavailableException("Database Temporarily Unavailable");
         return movieRepository.findAll();
     }
 
     @Override
-    public Movie updateMovie(Movie movie) /*throws MovieDoesNotExistException*/ {
-      /*  if (movieRepository.existsById(movie.getId()))
-            throw new MovieDoesNotExistException("Movie Not Found");*/
+    public Movie updateMovie(Movie movie) throws MovieDoesNotExistException {
+        if (movieRepository.existsById(movie.getId()))
+            throw new MovieDoesNotExistException("Movie Not Found");
         Movie updatedMovie = movieRepository.save(movie);
         return updatedMovie;
     }
 
     @Override
-    public String deleteMovie(int movieId) /*throws MovieDoesNotExistException*/ {
-//        if (!movieRepository.existsById(movieId)) {
-//            throw new MovieDoesNotExistException("Movie Does Not Exist");
-//        }
+    public String deleteMovie(int movieId) throws MovieDoesNotExistException {
+        if (!movieRepository.existsById(movieId)) {
+            throw new MovieDoesNotExistException("Movie Does Not Exist");
+            }
         movieRepository.deleteById(movieId);
         return "Movie deleted";
     }
